add: arquivos das funções de produtos
