+++ main.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 #include <stdlib.h>
-
+#include "moduloProduto/funcoesProduto.h"
 void menuSobre(void);
 char menuPrincipal(void);
 
-char menuProduto(void);
+//char menuProduto(void);
 void telaCadastroProduto(void);
 void telaPesquisarProduto(void);
 void telaApagarProduto(void);
@@ -44,9 +44,6 @@ void telaConfirmacao(void);
 
 int main(void) {
   char opcao;
-  // char opcaoTipoCliente;
-  // char opcaoAtualizarClientePF;
-  // char opcaoAtualizarClientePJ;
 
   menuSobre();
   
@@ -274,182 +271,6 @@ char menuPrincipal(void){
   scanf("%c",&opcao);
   getchar();
   printf("\n///            = = = = = = = = = = = = = = = = = = = =                      ///\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("\n");
-  printf("\t\t\t>>> Tecle <ENTER> para continuar...\n");
-  getchar();
-  return opcao;
-}
-
-char menuProduto(void){
-  char opcao;
-  system("clear");
-  printf("\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("///                                                                         ///\n");
-  printf("///              = = = = = = = = = = = = = = = = = = = =                    ///\n");
-  printf("///              =                                     =                    ///\n");
-  printf("///              =          Menu de Produtos           =                    ///\n");
-  printf("///              =                                     =                    ///\n");
-  printf("///              = = = = = = = = = = = = = = = = = = = =                    ///\n");
-  printf("///                                                                         ///\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("///                                                                         ///\n");
-  printf("///               1. Cadastrar um novo Produto                              ///\n");
-  printf("///               2. Pesquisar dados sobre um Produto                       ///\n");
-  printf("///               3. Apagar um Produto do sistema                           ///\n");
-  printf("///               4. Atualizar dados de um Produto                          ///\n");
-  printf("///               0. Volta para a tela anterior                             ///\n");
-  printf("///                                                                         ///\n");
-  printf("///                                                                         ///\n");
-  printf("                    Entre com sua opção desejada:   ");
-  scanf("%c",&opcao);
-  getchar();
-  printf("\n///            = = = = = = = = = = = = = = = = = = = =                      ///\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("\n");
-  printf("\t\t\t>>> Tecle <ENTER> para continuar...\n");
-  getchar();
-  return opcao;
-}
-
-
-
-void telaCadastroProduto(void){
-  char codigo[15];
-  char nome[25];
-  char validade[11];
-  int quantidade;
-  float preco;
-  char descricao[50];
-  system("clear");
-  printf("\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("///                                                                         ///\n");
-  printf("///              = = = = = = = = = = = = = = = = = = = =                    ///\n");
-  printf("///              =                                     =                    ///\n");
-  printf("///              =       Cadastrar novo Produto        =                    ///\n");
-  printf("///              =                                     =                    ///\n");
-  printf("///              = = = = = = = = = = = = = = = = = = = =                    ///\n");
-  printf("///                                                                         ///\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("///                                                                         ///\n");
-  printf("                  Código:   ");
-  scanf("%[0-9]",codigo);
-  getchar();
-  printf("                  Nome:   ");
-  scanf("%[A-Z a-zÁÉÍÓÚÃáéíóúã]",nome);
-  getchar();
-  printf("                  Quantidade:   ");
-  scanf("%d",&quantidade);
-  getchar();
-  printf("                  Validade:   ");
-  scanf("%[0-9/]",validade);
-  getchar();
-  printf("                  Preço da Unidade:   ");
-  scanf("%f",&preco);
-  getchar();
-  printf("                  Descrição:   ");
-  scanf("%[A-Z a-z0-9]",descricao);
-  getchar();
-  printf("\n///                                                                         ///\n");
-  printf("///                                                                         ///\n");
-  printf("///             = = = = = = = = = = = = = = = = = = = = =                   ///\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("\n");
-  printf("\t\t\t>>> Tecle <ENTER> para continuar...\n");
-  getchar();
-}
-
-void telaPesquisarProduto(void){
-  char codigo[15];
-  system("clear");
-  printf("\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("///                                                                         ///\n");
-  printf("///               = = = = = = = = = = = = = = = = = = = =                   ///\n");
-  printf("///               =                                     =                   ///\n");
-  printf("///               =     Pesquisar Produto no Sistema    =                   ///\n");
-  printf("///               =                                     =                   ///\n");
-  printf("///               = = = = = = = = = = = = = = = = = = = =                   ///\n");
-  printf("///                                                                         ///\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("///                                                                         ///\n");
-  printf("///                                                                         ///\n");
-  printf("///              Entre com o Código do produto a ser pesquisado             ///\n");
-  printf("///                                                                         ///\n");
-  printf("///                                                                         ///\n");
-  printf("                    Código:   ");
-  scanf("%[0-9]",codigo);
-  getchar();
-  printf("\n///                                                                         ///\n");
-  printf("///                                                                         ///\n");
-  printf("///             = = = = = = = = = = = = = = = = = = = = = = =               ///\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("\n");
-  printf("\t\t\t>>> Tecle <ENTER> para continuar...\n");
-  getchar();
-}
-
-void telaApagarProduto(void){
-  char codigo[15];
-  system("clear");
-  printf("\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("///                                                                         ///\n");
-  printf("///               = = = = = = = = = = = = = = = = = = = =                   ///\n");
-  printf("///               =                                     =                   ///\n");
-  printf("///               =      Apagar Produto do Sistema      =                   ///\n");
-  printf("///               =                                     =                   ///\n");
-  printf("///               = = = = = = = = = = = = = = = = = = = =                   ///\n");
-  printf("///                                                                         ///\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("///                                                                         ///\n");
-  printf("///                                                                         ///\n");
-  printf("///              Entre com o Código do produto a ser apagado                ///\n");
-  printf("///                                                                         ///\n");
-  printf("///                                                                         ///\n");
-  printf("                    Código:   ");
-  scanf("%[0-9]",codigo);
-  getchar();
-  printf("\n///                                                                         ///\n");
-  printf("///                                                                         ///\n");
-  printf("///             = = = = = = = = = = = = = = = = = = = = = = =               ///\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("\n");
-  printf("\t\t\t>>> Tecle <ENTER> para continuar...\n");
-  getchar();
-}
-
-char telaAtualizarProduto(void){
-  char opcao;
-  system("clear");
-  printf("\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("///                                                                         ///\n");
-  printf("///              = = = = = = = = = = = = = = = = = = = =                    ///\n");
-  printf("///              =                                     =                    ///\n");
-  printf("///              =    Atualizar Produto no Sistema     =                    ///\n");
-  printf("///              =                                     =                    ///\n");
-  printf("///              = = = = = = = = = = = = = = = = = = = =                    ///\n");
-  printf("///                                                                         ///\n");
-  printf("///////////////////////////////////////////////////////////////////////////////\n");
-  printf("///                                                                         ///\n");
-  printf("///               Opções de Atualização Disponíveis:                        ///\n");
-  printf("///                                                                         ///\n");
-  printf("///                                                                         ///\n");
-  printf("///                a. Atualizar Nome                                        ///\n");
-  printf("///                b. Atualizar Quantidade                                  ///\n");
-  printf("///                c. Atualizar Validade                                    ///\n");
-  printf("///                d. Atualizar Preço da Unidade                            ///\n");
-  printf("///                e. Atualizar Descrição                                   ///\n");
-  printf("///                                                                         ///\n");
-  printf("///                                                                         ///\n");
-  printf("                  Entre com uma opção desejada de atualização:   ");
-  scanf("%[a-e]",&opcao);
-  getchar();
-  printf("\n///                                                                         ///\n");
-  printf("///            = = = = = = = = = = = = = = = = = = = = = = =                ///\n");
   printf("///////////////////////////////////////////////////////////////////////////////\n");
   printf("\n");
   printf("\t\t\t>>> Tecle <ENTER> para continuar...\n");
