+++ moduloProduto/funcoesProduto.h
@@ -1,0 +1,5 @@
+char menuProduto(void);
+void telaCadastroProduto(void);
+void telaPesquisarProduto(void);
+void telaApagarProduto(void);
+char telaAtualizarProduto(void);
